+++ client_test.go
@@ -398,3 +398,116 @@ func TestClientNewSessionInvalidSecondResponseDifferentChannel(t *testing.T) {
 	require.Nil(t, session)
 	require.Error(t, client.Close())
 }
+
+func TestNewSessionTimedOut(t *testing.T) {
+	endAck := make(chan struct{})
+	responder := func(req frames.FrameBody) ([]byte, error) {
+		switch req.(type) {
+		case *mocks.AMQPProto:
+			return []byte{'A', 'M', 'Q', 'P', 0, 1, 0, 0}, nil
+		case *frames.PerformOpen:
+			return mocks.PerformOpen("container")
+		case *frames.PerformBegin:
+			// swallow the frame so NewSession never gets an ack
+			return nil, nil
+		case *frames.PerformEnd:
+			close(endAck)
+			return mocks.PerformEnd(0, nil)
+		default:
+			return nil, fmt.Errorf("unhandled frame %T", req)
+		}
+	}
+	netConn := mocks.NewNetConn(responder)
+
+	client, err := New(netConn, nil)
+	require.NoError(t, err)
+	// fisrt session succeeds
+	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
+	session, err := client.NewSession(ctx, nil)
+	cancel()
+	require.ErrorIs(t, err, context.DeadlineExceeded)
+	require.Nil(t, session)
+
+	select {
+	case <-time.After(time.Second):
+		t.Fatal("didn't receive end ack")
+	case <-endAck:
+		// expected
+	}
+}
+
+func TestNewSessionWriteError(t *testing.T) {
+	endAck := make(chan struct{})
+	responder := func(req frames.FrameBody) ([]byte, error) {
+		switch req.(type) {
+		case *mocks.AMQPProto:
+			return []byte{'A', 'M', 'Q', 'P', 0, 1, 0, 0}, nil
+		case *frames.PerformOpen:
+			return mocks.PerformOpen("container")
+		case *frames.PerformBegin:
+			return nil, errors.New("write error")
+		case *frames.PerformEnd:
+			close(endAck)
+			return mocks.PerformEnd(0, nil)
+		default:
+			return nil, fmt.Errorf("unhandled frame %T", req)
+		}
+	}
+	netConn := mocks.NewNetConn(responder)
+
+	client, err := New(netConn, nil)
+	require.NoError(t, err)
+	// fisrt session succeeds
+	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
+	session, err := client.NewSession(ctx, nil)
+	cancel()
+	var connErr *ConnectionError
+	require.ErrorAs(t, err, &connErr)
+	require.Equal(t, "write error", connErr.Error())
+	require.Nil(t, session)
+
+	select {
+	case <-time.After(time.Second):
+		// expected
+	case <-endAck:
+		t.Fatal("unexpected ack")
+	}
+}
+
+func TestNewSessionTimedOutAckTimedOut(t *testing.T) {
+	endAck := make(chan struct{})
+	responder := func(req frames.FrameBody) ([]byte, error) {
+		switch req.(type) {
+		case *mocks.AMQPProto:
+			return []byte{'A', 'M', 'Q', 'P', 0, 1, 0, 0}, nil
+		case *frames.PerformOpen:
+			return mocks.PerformOpen("container")
+		case *frames.PerformBegin:
+			// swallow the frame so NewSession never gets an ack
+			return nil, nil
+		case *frames.PerformEnd:
+			close(endAck)
+			// swallow the frame so the closing goroutine never gets an ack
+			return nil, nil
+		default:
+			return nil, fmt.Errorf("unhandled frame %T", req)
+		}
+	}
+	netConn := mocks.NewNetConn(responder)
+
+	client, err := New(netConn, nil)
+	require.NoError(t, err)
+	// fisrt session succeeds
+	ctx, cancel := context.WithTimeout(context.Background(), 1*time.Second)
+	session, err := client.NewSession(ctx, nil)
+	cancel()
+	require.ErrorIs(t, err, context.DeadlineExceeded)
+	require.Nil(t, session)
+
+	select {
+	case <-time.After(time.Second):
+		t.Fatal("didn't receive end ack")
+	case <-endAck:
+		// expected
+	}
+}