+++ session.go
@@ -167,15 +167,7 @@ func (s *Session) NewSender(ctx context.Context, target string, opts *SenderOpti
 
 func (s *Session) mux(remoteBegin *frames.PerformBegin) {
 	defer func() {
-		// clean up session record in conn.mux()
-		select {
-		case <-s.rx:
-			// discard any incoming frames to keep conn mux unblocked
-		case s.conn.DelSession <- s:
-			// successfully deleted session
-		case <-s.conn.Done:
-			s.err = s.conn.Err()
-		}
+		s.conn.DeleteSession(s)
 		if s.err == nil {
 			s.err = ErrSessionClosed
 		}